Instalando y configurando redux.
- Se crearon los directorios reducers, store, types.
- Se creo el primer reducer para manejar el login.
- Se creo el componente AppRouter para manejar el sistema de rutas.
- Se instalaron las dependencias redux y react-redux.
