+++ src/store/store.js
@@ -1,0 +1,8 @@
+import { createStore, combineReducers } from "redux";
+import { authReducer } from "../reducers/authReducer";
+
+const reducers = combineReducers({
+    auth: authReducer
+})
+
+export const store = createStore(reducers)
