+++ src/types/types.js
@@ -1,0 +1,6 @@
+
+
+export const types = {
+    login: '[Auth] Login',
+    logout: '[Auth] Logout'
+}
